plantuml2statemachine/requirements.txt
pydot==4.0.0
pyparsing==3.2.3
python-statemachine==2.5.0
